backend/collaboration_service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-socketio==5.11.0
python-engineio==4.9.0
bidict==0.23.1
aiohttp==3.9.3
asyncio==3.4.3
python-dotenv==1.0.1
pydantic-settings==2.2.1
redis==5.0.1